Produce distroless minimized container images
This change produces a minimal container image (~25MB) containing only
the entrypoint bash (static) executable, the split-sync or split-proxy
executable, and the minimal Debian files needed for these to execute.

The dependency on the `tr` application was removed from functions.sh
by using bash variable parameter expansion instead.

We use a static bash executable to minimize libraries needed to
include in the image.  Bash executes only briefly in the entrypoint
which then exec's into the application.

split-sync and split-proxy are dynamically linked with libc because
the go net library requires it.  Therefore, libc is copied into the
final image too.

ca-certificates are copied into the image so TLS connections to
split.io can be validated.

The last build stage is used to produce a single-layer final
container.