Add healthcheck script
split-synchronizer doesn’t come with a health check program in the
container callable by an ECS Task Definition.  We can add one using
bash only, which is already present in the image, so we don't have to
add something large such as curl.  The executable includes a GET
/health/application endpoint on port 3010 which returns HTTP 200 OK on
healthy, HTTP 500 on unhealthy.

This script performs the minimal HTTP protocol call, placing the GET
call and reading the first returned line for the status code, ignoring
all other headers and body.